% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODB_functions.R
\name{modb.clean.file}
\alias{modb.clean.file}
\title{Clean all files and folders inside a directory}
\usage{
modb.clean.file(idx_dir)
}
\arguments{
\item{idx_dir}{Character. Directory to clean.}
}
\value{
Invisibly TRUE.
}
\description{
Create the directory if it doesn't exist. And make sure it being cleaned
}
\examples{
modb.clean.file("IndexedData")
modb.clean.file("Output")

}
