% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Start_Shiny_UI.R
\name{MODB.Run}
\alias{MODB.Run}
\title{Run Shiny UI for the Genomic DB Browser}
\usage{
MODB.Run(use_current = FALSE)
}
\arguments{
\item{use_current}{Logical.
\itemize{
\item \code{FALSE} (default): run packaged app
\item \code{TRUE}: copy only app.R and MODB_functions.r into current directory,
then run app locally
}}
}
\value{
Invisibly returns the directory used for launching the app.
}
\description{
Launches the Shiny-based Genomic DB Browser.
\itemize{
\item Default: Runs the packaged UI located inside the installed
MODBrowser package.
\item Development mode (\code{use_current = TRUE}): need to run xiaopei.input.all() first to build dataset index.
It Copies only two files into the
\strong{current working directory}:
\itemize{
\item app.R
\item MODB_functions.r
}

Then runs the Shiny app from the working directory.

This allows users to modify UI/server code locally without altering the package installation
and without creating any additional folders.
Under development mode: since the datasets are updated,
it's required to run "Rebuild EVERYTHING" at Administrator Operations(left-bottom side) after launched ShinyUI
}
}
